Edit product rating field and add products

+++ products/models.py
@@ -25,7 +25,7 @@ class Product(models.Model):
     name = models.CharField(max_length=256)
     description = models.TextField()
     price = models.DecimalField(max_digits=5, decimal_places=2)
-    rating = models.IntegerField(null=True, blank=True)
+    rating = models.DecimalField(max_digits=5, decimal_places=2, null=True, blank=True)
     image_url = models.URLField(max_length=1024, null=True, blank=True)
     image = models.ImageField(null=True, blank=True)
